now test and train dataset can in the same path

+++ pst.py
@@ -164,7 +164,7 @@ for i in range(len(classeslist)):#all classes
         fixTrainORTestTxt(classes_train,num_tuple_train)        
     if os.path.isfile(classes_val):
         fixTrainORTestTxt(classes_val,num_tuple_val) 
-#remove jpeg,segmentationClass,segmentationObject
+#remove jpeg,segmentationClass,segmentationObject,Annotation
 DirTuple=(VOC_JPEGImages,VOC_SegmentationClass,VOC_SegmentationObject,VOC_Annotation)
 for i in range(len(DirTuple)):
     list = os.listdir(DirTuple[i])
@@ -173,11 +173,11 @@ for i in range(len(DirTuple)):
         if os.path.isfile(path):
             file_name_list=list[j].split('.')
             if len(num_set_test) is not 0:
-                if file_name_list[0] not in num_tuple_test:
+                if file_name_list[0] not in num_tuple_test and file_name_list[0] not in num_tuple_trainval:
                     if os.path.isfile(path):
                         os.remove(path)
             if len(num_set_trainval) is not 0:
-                if file_name_list[0] not in num_tuple_trainval:
+                if file_name_list[0] not in num_tuple_trainval and file_name_list[0] not in num_tuple_test:
                     if os.path.isfile(path):
                         os.remove(path)
 #edit xml
